[ARC-525] Remove unnecessary packages for fundamental types
branch: feature/ARC-525-report-results
