+++ src/mpyl/stages/discovery.py
@@ -5,7 +5,7 @@ output artifact."""
 from ..project import Project
 from ..projects.find import load_projects
 from ..repo import History, Repository
-from ..stage import Stage
+from .. import Stage
 from ..steps.models import Output
