+++ src/mpyl/steps/deploy/echo.py
@@ -5,7 +5,7 @@ from logging import Logger
 from ..step import Step
 
 from ..models import Meta, Input, Output, ArtifactType
-from ...stage import Stage
+from ... import Stage
 
 
 class DeployEcho(Step):
