+++ src/mpyl/steps/models.py
@@ -8,8 +8,8 @@ from typing import Optional, Dict
 from ruamel.yaml import YAML, yaml_object  # type: ignore
 
 from ..project import Project
-from ..stage import Stage
-from ..target import Target
+from .. import Stage
+from .. import Target
 
 yaml = YAML()
