+++ tests/test_project.py
@@ -4,7 +4,7 @@ import pytest
 from jsonschema import ValidationError
 
 from src.mpyl.project import load_project
-from src.mpyl.target import Target
+from src.mpyl import Target
 from tests import root_test_path
